requirements.txt
flask>=2.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
